fixing header files, refactored node
